pyproject.toml
[project]
name = "businessmanagementsystem"
version = "0.1.0"
description = ""
authors = [
    {name = "Vadim",email = "marakushin.vadim@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi[standart] (>=0.135.2,<0.136.0)",
    "black (>=26.3.1,<27.0.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "fastapi-admin (>=1.0.4,<2.0.0)",
    "fastapi-users (>=15.0.5,<16.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "fastapi-users-db-sqlalchemy (>=7.0.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
